Move shader load functions to their own files
