+++ systems/triangl_render.cpp
@@ -3,74 +3,10 @@
 //
 
 #include <iostream>
-#include <fstream>
-#include <sstream>
-#include <string>
 #include <algorithm>
+
 #include "triangl_render.h"
-
-// Loads contents of shader file into a string
-std::string readShaderFile(const std::string& filePath) {
-    std::ifstream file(filePath);
-    if (!file.is_open()) {
-        std::cerr << "Failed to open file: " << filePath << std::endl;
-        return "";
-    }
-
-    std::stringstream buffer;
-    buffer << file.rdbuf();
-    return buffer.str();
-}
-
-// Compile the shader and error check it
-GLuint compileShader(const std::string& shaderSource, GLenum shaderType) {
-    GLuint shader = glCreateShader(shaderType);
-    const char* sourceCStr = shaderSource.c_str();
-    glShaderSource(shader, 1, &sourceCStr, nullptr);
-    glCompileShader(shader);
-
-    GLint success;
-    glGetShaderiv(shader, GL_COMPILE_STATUS, &success);
-    if (!success) {
-        GLchar infoLog[512];
-        glGetShaderInfoLog(shader, 512, nullptr, infoLog);
-        std::cerr << "Error compiling shader: " << infoLog << std::endl;
-        glDeleteShader(shader);
-        return 0;
-    }
-
-    return shader;
-}
-
-// Create the shader program
-GLuint createShaderProgram(const std::string& vertexShaderSource, const std::string& fragmentShaderSource) {
-    GLuint vertexShader = compileShader(vertexShaderSource, GL_VERTEX_SHADER);
-    GLuint fragmentShader = compileShader(fragmentShaderSource, GL_FRAGMENT_SHADER);
-
-    if (vertexShader == 0 || fragmentShader == 0) {
-        return 0;
-    }
-
-    GLuint shaderProgram = glCreateProgram();
-    glAttachShader(shaderProgram, vertexShader);
-    glAttachShader(shaderProgram, fragmentShader);
-    glLinkProgram(shaderProgram);
-
-    GLint success;
-    glGetProgramiv(shaderProgram, GL_LINK_STATUS, &success);
-    if (!success) {
-        GLchar infoLog[512];
-        glGetProgramInfoLog(shaderProgram, 512, nullptr, infoLog);
-        std::cerr << "Error linking shader program: " << infoLog << std::endl;
-        glDeleteProgram(shaderProgram);
-        return 0;
-    }
-
-    glDeleteShader(vertexShader);
-    glDeleteShader(fragmentShader);
-
-    return shaderProgram;
-}
+#include "../shaders.h"
 
 triangl_render::triangl_render() {
